+++ src/aux-common/runtime/AuxRuntime.ts
@@ -156,9 +156,10 @@ export class AuxRuntime
         this._onActions = new Subject();
         this._onErrors = new Subject();
 
-        this._sub = new Subscription(() => {
+        let sub = (this._sub = new Subscription(() => {
             this._globalContext.cancelAllBotTimers();
-        });
+        }));
+        sub.add(this._globalContext.startAnimationLoop());
     }
 
     getShoutTimers(): { [shout: string]: number } {
@@ -288,7 +289,7 @@ export class AuxRuntime
         let rejected = false;
         for (let i = 0; i < result.actions.length; i++) {
             const a = result.actions[i];
-            if (a.type === 'reject' && a.action === action) {
+            if (a.type === 'reject' && a.actions.indexOf(action) >= 0) {
                 rejected = true;
                 result.actions.splice(i, 1);
                 break;